hago que el protocolo del server reciba el client_id del client_handler para crear el comando, y modifico el game_id de games_handler de un int* a un uint8_t&

+++ src/server/runnable_commands/start_moving_runnable.h
@@ -7,7 +7,7 @@
 class RunnableStartMoving : public RunnableCommandGame {
 
 public:
-  explicit RunnableStartMoving(int clt_id, Socket &skt, bool *was_closed)
+  explicit RunnableStartMoving(uint8_t clt_id, Socket &skt, bool *was_closed)
       : RunnableCommandGame(
             std::make_shared<StartMoving>(clt_id, skt, was_closed)) {}
